Python-interface: Loading spin-polarized quantities (#170)
* Python-interface: Loading spin-polarized quantities

This commit fixes a segmentation fault that occurred when trying to
load certain result properties in spin-polarized calculations. The
Python-interface is extended to provide convenient access to
spin-polarized quantities such as densities, orbital coefficients,
occupation numbers, and orbital energies.

Signed-off-by: Yannick Lemke <yannick.lemke@posteo.de>

* Python-interface: Update docstring for `Result.get()`

Signed-off-by: Yannick Lemke <yannick.lemke@posteo.de>

* Python-interface: Fix ordering of loaded matrices

The matrices in the Python-interface are defined to be in row-major
ordering, while the data written by Fortran is in column-major ordering.
This leads to orbital coefficients having the MO index as their first
axis, which is misleading. This commit swaps the axes of the matrices to
ensure interoperability.

Signed-off-by: Yannick Lemke <yannick.lemke@posteo.de>

* Fix `natoms` and `charges` returned by the Python interface

For spin-polarized calculations, `res.get("natoms")` currently
returns twice the number of atoms, and `res.get("charges")` returns
an array of shape `(2 * natoms,)` with the second half being filled
with zeros. This commit fixes these issues.

Signed-off-by: Yannick Lemke <yannick.lemke@posteo.de>

* Add more spin-polarized tests to the Python interface

Add more tests for spin-polarized quantities to the Python interface to
ensure proper dimensions and values of the underlying arrays.

Signed-off-by: Yannick Lemke <yannick.lemke@posteo.de>

* Fix code block rendering (#171)

* Add missing import (#172)

* Add toml read and write functionality to double dictionary. (#158)

* fix test suite to capture errors
fix remove functionality
add toml read and write

* change condition in comparison cycle

* brake line, correct test comparison

* cover more lines with tests

* fix gnu compiler issues

* merge toml-r-w-changes from prev branch

* add toml-r-w func

* use .not.

* remove toml feature

* remove module loads and whitespace

* remove extra end type

* fix toml output

---------

Co-authored-by: Pit Steinbach <pit.steinbach@rwth-aachen.de>
Co-authored-by: Pit Steinbach <steinbach@pc.rwth-aachen.de>

* interface.py: improve docstring table readability

Signed-off-by: Yannick Lemke <yannick.lemke@posteo.de>

* API tests: additional checks for `get_number_of_spins`

Signed-off-by: Yannick Lemke <yannick.lemke@posteo.de>

---------

Signed-off-by: Yannick Lemke <yannick.lemke@posteo.de>
Co-authored-by: Marvin Friede <51965259+marvinfriede@users.noreply.github.com>
Co-authored-by: Pit Steinbach <73583910+pitsteinbach@users.noreply.github.com>
Co-authored-by: Pit Steinbach <pit.steinbach@rwth-aachen.de>
Co-authored-by: Pit Steinbach <steinbach@pc.rwth-aachen.de>